requirements.txt
streamlit
numpy
pandas
matplotlib
seaborn
plotly
mysql-connector-python
python-dotenv
extra-streamlit-components
bcrypt